user_service/requirements.txt
# user_service/requirements.txt

django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn